chore: use latest amqp version

+++ src/api/src/utils/config.rs
@@ -10,7 +10,7 @@ pub struct Config {
   pub exchange_name: String,
   pub queue_name: String,
   pub routing_key: String,
-  pub retry_ttl: u16,
+  pub retry_ttl: u32,
   pub enc_key: Vec<u8>,
 }
 
@@ -26,7 +26,7 @@ impl Config {
         exchange_name: env::var("EXCHANGE_NAME").unwrap(),
         queue_name: env::var("QUEUE_NAME").unwrap(),
         routing_key: env::var("ROUTING_KEY").unwrap(),
-        retry_ttl: env::var("RETRY_TTL").unwrap().parse::<u16>().unwrap(),
+        retry_ttl: env::var("RETRY_TTL").unwrap().parse::<u32>().unwrap(),
         enc_key: env::var("ENC_KEY").unwrap().as_bytes().to_vec(),
       }
     )
